api/requirements.txt
Flask
Gunicorn
pymysql
SQLAlchemy
Flask-hCaptcha
python-dotenv
cryptography